16/4/20 users테이블과 posts테이블 1:N 관계 형성

+++ src/main/java/com/pilot/service/UserService.java
@@ -16,7 +16,7 @@ public class UserService {
 	@Autowired
 	private UserRepository userRepository;
 
-	public User findOne(Long id){
+	public User findOne(Integer id){
 		return userRepository.findOne(id);
 	}
